+++ lib/src/dartdap/core/ldap_util.dart
@@ -1,7 +1,6 @@
-part of dartdap;
 
 // todo: why a class? Make a top level functions?
-class _LdapUtil {
+class LdapUtil {
   /**
    * Escape an ldap string used in a search filter.
    * The LDAP spec requires *,),(,\ and null to be escaped.