+++ lib/src/dartdap/core/modification.dart
@@ -1,4 +1,3 @@
-part of dartdap;
 
 /**
  * An LDAP modification operation type.