+++ lib/src/dartdap/core/search_result.dart
@@ -1,5 +1,5 @@
-part of dartdap;
-
+import 'ldap_result.dart';
+import '../control/control.dart';
 /// Results from a search request.
 ///
 /// This object is produced by the [LdapConnection.search] method.