+++ lib/src/dartdap/core/search_scope.dart
@@ -1,4 +1,3 @@
-part of dartdap;
 
 class SearchScope {
   static const int BASE_LEVEL = 0;